renaming project to HayesEQ
