+++ HayesEQ.jucer
@@ -1,0 +1,113 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<JUCERPROJECT name="HayesEQ" companyName="EmmetSoft" version="1.0.0" displaySplashScreen="1"
+              defines="PIP_JUCE_EXAMPLES_DIRECTORY=L2hvbWUvZW1tZXQvRG9jdW1lbnRzL0pVQ0UvZXhhbXBsZXM="
+              projectType="audioplug" useAppConfig="0" addUsingNamespaceToJuceHeader="1"
+              id="pRVjbr" jucerFormatVersion="1" pluginFormats="buildAU,buildLV2,buildStandalone,buildVST3">
+  <MAINGROUP id="BE1GVm" name="HayesEQ">
+    <GROUP id="{CE2A1EFF-709F-E63D-44FF-CA2F032D66D5}" name="Source">
+      <FILE id="N3iDqw" name="Utilities.h" compile="0" resource="0" file="Source/Utilities.h"/>
+      <FILE id="yLzhSX" name="SpectrumPlotComponent.h" compile="0" resource="0"
+            file="Source/SpectrumPlotComponent.h"/>
+      <FILE id="fPSiMo" name="CentredLabel.h" compile="0" resource="0" file="Source/CentredLabel.h"/>
+      <FILE id="cCUZyb" name="IIRFilterBandComponent.h" compile="0" resource="0"
+            file="Source/IIRFilterBandComponent.h"/>
+      <FILE id="izaYWB" name="LogSlider.h" compile="0" resource="0" file="Source/LogSlider.h"/>
+      <FILE id="tVxFY5" name="CustomLookAndFeel.h" compile="0" resource="0"
+            file="Source/CustomLookAndFeel.h"/>
+      <FILE id="vT4KGJ" name="HayesEQAudioProcessorEditor.cpp" compile="1"
+            resource="0" file="Source/HayesEQAudioProcessorEditor.cpp"/>
+      <FILE id="SnLJHA" name="HayesEQAudioProcessorEditor.h" compile="0"
+            resource="0" file="Source/HayesEQAudioProcessorEditor.h"/>
+      <FILE id="izy5K0" name="HayesEQAudioProcessor.cpp" compile="1" resource="0"
+            file="Source/HayesEQAudioProcessor.cpp"/>
+      <FILE id="d0Q1Qr" name="HayesEQAudioProcessor.h" compile="0" resource="0"
+            file="Source/HayesEQAudioProcessor.h"/>
+    </GROUP>
+    <FILE id="eIH07U" name="bgfile.jpg" compile="0" resource="1" file="Images/bgfile.jpg"/>
+  </MAINGROUP>
+  <MODULES>
+    <MODULE id="juce_audio_basics" showAllCode="1" useLocalCopy="0" useGlobalPath="1"/>
+    <MODULE id="juce_audio_devices" showAllCode="1" useLocalCopy="0" useGlobalPath="1"/>
+    <MODULE id="juce_audio_formats" showAllCode="1" useLocalCopy="0" useGlobalPath="1"/>
+    <MODULE id="juce_audio_plugin_client" showAllCode="1" useLocalCopy="0"
+            useGlobalPath="1"/>
+    <MODULE id="juce_audio_processors" showAllCode="1" useLocalCopy="0" useGlobalPath="1"/>
+    <MODULE id="juce_audio_utils" showAllCode="1" useLocalCopy="0" useGlobalPath="1"/>
+    <MODULE id="juce_core" showAllCode="1" useLocalCopy="0" useGlobalPath="1"/>
+    <MODULE id="juce_data_structures" showAllCode="1" useLocalCopy="0" useGlobalPath="1"/>
+    <MODULE id="juce_dsp" showAllCode="1" useLocalCopy="0" useGlobalPath="1"/>
+    <MODULE id="juce_events" showAllCode="1" useLocalCopy="0" useGlobalPath="1"/>
+    <MODULE id="juce_graphics" showAllCode="1" useLocalCopy="0" useGlobalPath="1"/>
+    <MODULE id="juce_gui_basics" showAllCode="1" useLocalCopy="0" useGlobalPath="1"/>
+    <MODULE id="juce_gui_extra" showAllCode="1" useLocalCopy="0" useGlobalPath="1"/>
+  </MODULES>
+  <EXPORTFORMATS>
+    <XCODE_MAC targetFolder="Builds/MacOSX">
+      <CONFIGURATIONS>
+        <CONFIGURATION name="Debug" isDebug="1" optimisation="1" targetName="HayesEQ"/>
+        <CONFIGURATION name="Release" isDebug="0" optimisation="3" targetName="HayesEQ"/>
+      </CONFIGURATIONS>
+      <MODULEPATHS>
+        <MODULEPATH id="juce_audio_basics" path=""/>
+        <MODULEPATH id="juce_audio_devices" path=""/>
+        <MODULEPATH id="juce_audio_formats" path=""/>
+        <MODULEPATH id="juce_audio_processors" path=""/>
+        <MODULEPATH id="juce_audio_utils" path=""/>
+        <MODULEPATH id="juce_core" path=""/>
+        <MODULEPATH id="juce_data_structures" path=""/>
+        <MODULEPATH id="juce_dsp" path=""/>
+        <MODULEPATH id="juce_events" path=""/>
+        <MODULEPATH id="juce_graphics" path=""/>
+        <MODULEPATH id="juce_gui_basics" path=""/>
+        <MODULEPATH id="juce_gui_extra" path=""/>
+        <MODULEPATH id="juce_audio_plugin_client" path="../JUCE/modules"/>
+      </MODULEPATHS>
+    </XCODE_MAC>
+    <LINUX_MAKE targetFolder="Builds/LinuxMakefile">
+      <CONFIGURATIONS>
+        <CONFIGURATION name="Debug" isDebug="1" optimisation="1" targetName="HayesEQ"
+                       headerPath="/usr/include/freetype2&#10;/usr/include/gtk-3.0&#10;/usr/local/include/gtk-3.0/gdk&#10;/usr/local/include/gdk-pixbuf-2.0&#10;/usr/local/include/atk-1.0&#10;/usr/local/lib/gtk-3.0/include&#10;/usr/include/glib-2.0&#10;/usr/lib/x86_64-linux-gnu/glib-2.0/include&#10;/usr/include/cairo&#10;/usr/include/pango-1.0&#10;/usr/include/harfbuzz&#10;/usr/include/webkitgtk-4.0&#10;/usr/include/libsoup-2.4"/>
+        <CONFIGURATION name="Release" isDebug="0" optimisation="3" targetName="HayesEQ"
+                       headerPath="/usr/include/freetype2&#10;/usr/include/gtk-3.0&#10;/usr/local/include/gtk-3.0/gdk&#10;/usr/local/include/gdk-pixbuf-2.0&#10;/usr/local/include/atk-1.0&#10;/usr/local/lib/gtk-3.0/include&#10;/usr/include/glib-2.0&#10;/usr/lib/x86_64-linux-gnu/glib-2.0/include&#10;/usr/include/cairo&#10;/usr/include/pango-1.0&#10;/usr/include/harfbuzz&#10;/usr/include/webkitgtk-4.0&#10;/usr/include/libsoup-2.4"/>
+      </CONFIGURATIONS>
+      <MODULEPATHS>
+        <MODULEPATH id="juce_audio_basics" path=""/>
+        <MODULEPATH id="juce_audio_devices" path=""/>
+        <MODULEPATH id="juce_audio_formats" path=""/>
+        <MODULEPATH id="juce_audio_processors" path=""/>
+        <MODULEPATH id="juce_audio_utils" path=""/>
+        <MODULEPATH id="juce_core" path=""/>
+        <MODULEPATH id="juce_data_structures" path=""/>
+        <MODULEPATH id="juce_dsp" path=""/>
+        <MODULEPATH id="juce_events" path=""/>
+        <MODULEPATH id="juce_graphics" path=""/>
+        <MODULEPATH id="juce_gui_basics" path=""/>
+        <MODULEPATH id="juce_gui_extra" path=""/>
+        <MODULEPATH id="juce_audio_plugin_client" path="../JUCE/modules"/>
+      </MODULEPATHS>
+    </LINUX_MAKE>
+    <VS2019 targetFolder="Builds/VisualStudio2019">
+      <CONFIGURATIONS>
+        <CONFIGURATION isDebug="1" name="Debug" targetName="HayesEQ"/>
+        <CONFIGURATION isDebug="0" name="Release" targetName="HayesEQ"/>
+      </CONFIGURATIONS>
+      <MODULEPATHS>
+        <MODULEPATH id="juce_gui_extra" path="../../juce"/>
+        <MODULEPATH id="juce_gui_basics" path="../../juce"/>
+        <MODULEPATH id="juce_graphics" path="../../juce"/>
+        <MODULEPATH id="juce_events" path="../../juce"/>
+        <MODULEPATH id="juce_dsp" path="../../juce"/>
+        <MODULEPATH id="juce_data_structures" path="../../juce"/>
+        <MODULEPATH id="juce_core" path="../../juce"/>
+        <MODULEPATH id="juce_audio_utils" path="../../juce"/>
+        <MODULEPATH id="juce_audio_processors" path="../../juce"/>
+        <MODULEPATH id="juce_audio_plugin_client" path="../../juce"/>
+        <MODULEPATH id="juce_audio_formats" path="../../juce"/>
+        <MODULEPATH id="juce_audio_devices" path="../../juce"/>
+        <MODULEPATH id="juce_audio_basics" path="../../juce"/>
+      </MODULEPATHS>
+    </VS2019>
+  </EXPORTFORMATS>
+  <JUCEOPTIONS JUCE_STRICT_REFCOUNTEDPOINTER="1" JUCE_VST3_CAN_REPLACE_VST2="0"/>
+</JUCERPROJECT>
