+++ JuceLibraryCode/JuceHeader.h
@@ -47,7 +47,7 @@
 #if ! JUCE_DONT_DECLARE_PROJECTINFO
 namespace ProjectInfo
 {
-    const char* const  projectName    = "IIRFilter";
+    const char* const  projectName    = "HayesEQ";
     const char* const  companyName    = "EmmetSoft";
     const char* const  versionString  = "1.0.0";
     const int          versionNumber  = 0x10000;
