+++ JuceLibraryCode/JuceLV2Defines.h
@@ -1,5 +1,5 @@
 #pragma once
 
 #ifndef JucePlugin_LV2URI
- #define JucePlugin_LV2URI "www.EmmetSoft.com/plugins/IIRFilter"
+ #define JucePlugin_LV2URI "www.EmmetSoft.com/plugins/HayesEQ"
 #endif
