+++ JuceLibraryCode/JucePluginDefines.h
@@ -41,10 +41,10 @@
  #define JucePlugin_Enable_ARA             0
 #endif
 #ifndef  JucePlugin_Name
- #define JucePlugin_Name                   "IIRFilter"
+ #define JucePlugin_Name                   "HayesEQ"
 #endif
 #ifndef  JucePlugin_Desc
- #define JucePlugin_Desc                   "IIRFilter"
+ #define JucePlugin_Desc                   "HayesEQ"
 #endif
 #ifndef  JucePlugin_Manufacturer
  #define JucePlugin_Manufacturer           "EmmetSoft"
@@ -101,19 +101,19 @@
  #define JucePlugin_AUSubType              JucePlugin_PluginCode
 #endif
 #ifndef  JucePlugin_AUExportPrefix
- #define JucePlugin_AUExportPrefix         IIRFilterAU
+ #define JucePlugin_AUExportPrefix         HayesEQAU
 #endif
 #ifndef  JucePlugin_AUExportPrefixQuoted
- #define JucePlugin_AUExportPrefixQuoted   "IIRFilterAU"
+ #define JucePlugin_AUExportPrefixQuoted   "HayesEQAU"
 #endif
 #ifndef  JucePlugin_AUManufacturerCode
  #define JucePlugin_AUManufacturerCode     JucePlugin_ManufacturerCode
 #endif
 #ifndef  JucePlugin_CFBundleIdentifier
- #define JucePlugin_CFBundleIdentifier     com.EmmetSoft.IIRFilter
+ #define JucePlugin_CFBundleIdentifier     com.EmmetSoft.HayesEQ
 #endif
 #ifndef  JucePlugin_AAXIdentifier
- #define JucePlugin_AAXIdentifier          com.EmmetSoft.IIRFilter
+ #define JucePlugin_AAXIdentifier          com.EmmetSoft.HayesEQ
 #endif
 #ifndef  JucePlugin_AAXManufacturerCode
  #define JucePlugin_AAXManufacturerCode    JucePlugin_ManufacturerCode
@@ -137,7 +137,7 @@
  #define JucePlugin_IAASubType             JucePlugin_PluginCode
 #endif
 #ifndef  JucePlugin_IAAName
- #define JucePlugin_IAAName                "EmmetSoft: IIRFilter"
+ #define JucePlugin_IAAName                "EmmetSoft: HayesEQ"
 #endif
 #ifndef  JucePlugin_VSTNumMidiInputs
  #define JucePlugin_VSTNumMidiInputs       16
@@ -152,10 +152,10 @@
  #define JucePlugin_ARATransformationFlags  0
 #endif
 #ifndef  JucePlugin_ARAFactoryID
- #define JucePlugin_ARAFactoryID           "com.EmmetSoft.IIRFilter.factory"
+ #define JucePlugin_ARAFactoryID           "com.EmmetSoft.HayesEQ.factory"
 #endif
 #ifndef  JucePlugin_ARADocumentArchiveID
- #define JucePlugin_ARADocumentArchiveID   "com.EmmetSoft.IIRFilter.aradocumentarchive.1.0.0"
+ #define JucePlugin_ARADocumentArchiveID   "com.EmmetSoft.HayesEQ.aradocumentarchive.1.0.0"
 #endif
 #ifndef  JucePlugin_ARACompatibleArchiveIDs
  #define JucePlugin_ARACompatibleArchiveIDs  ""
